modules: move LmodConfiguration defs into BaseConfiguration
Move hierarchical-specific methods of LmodConfiguration class into
generic BaseConfiguration to allow hierarchical setup also for Tcl.

* Also move related global methods and error classes of
  spack.modules.lmod into spack.modules.common.
* Add "module_system" method to BaseConfiguration class that returns
  name of currently used module system (either "tcl" or "lmod").
* Add argument "module_system" to guess_core_compilers global function.
* Update "requires" method of BaseConfiguration class to return an empty
  dictionnary if hierarchical mode is disabled.
* Adapt "default_projections" configuration depending if hierarchical
  mode is enabled or not.
* Make "provides" and "requires" methods return an empty dictionary if
  hierarchical mode is disabled.
